refactoring of lunch service
